Rename helper js files
