+++ frontend/src/components/CartProductCard.jsx
@@ -1,5 +1,5 @@
 import { Minus, Plus } from "lucide-react";
-import { apiDomain } from "../helpers/api-domain";
+import { apiDomain } from "../helpers/apiDomain";
 import { Link } from "react-router-dom";
 import { useCart } from "../hooks/useCart";
