+++ frontend/test/helpers/hide-credit-card-number.test.js
@@ -1,5 +1,5 @@
 import { expect, test } from "vitest";
-import hideCreditCardNumber from "../../src/helpers/hide-credit-card-number";
+import hideCreditCardNumber from "../../src/helpers/hideCreditCardNumber";
 
 test("Obfuscate credit card number with stars (*)", () => {
   expect(hideCreditCardNumber("1234 5678 9012 3456")).toBe("**** **** **** 3456");
